storage: extract Key to tikv_misc
