+++ components/raftstore/src/coprocessor/mod.rs
@@ -1,6 +1,4 @@
-use engine::rocks::DB;
 use kvproto::metapb::Region;
-use kvproto::pdpb::CheckPolicy;
 use kvproto::raft_cmdpb::{AdminRequest, AdminResponse, Request, Response};
 use protobuf::RepeatedField;
 use raft::StateRole;
