+++ components/tikv_misc/Cargo.toml
@@ -9,6 +9,7 @@ byteorder = "1.2"
 engine = { path = "../engine" }
 futures = "0.1"
 grpcio = { version = "0.4" }
+hex = "0.3"
 kvproto = { git = "https://github.com/pingcap/kvproto.git" }
 protobuf = "~2.0"
 quick-error = "1.2.2"
